+++ driver-core/src/test/java/com/datastax/driver/core/AuthenticationTest.java
@@ -16,6 +16,7 @@
 package com.datastax.driver.core;
 
 import java.net.InetAddress;
+import java.net.InetSocketAddress;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
@@ -67,7 +68,7 @@ public class AuthenticationTest {
                                                 .connect();
         } catch (NoHostAvailableException e) {
 
-            for (Map.Entry<InetAddress, Throwable> entry : e.getErrors().entrySet())
+            for (Map.Entry<InetSocketAddress, Throwable> entry : e.getErrors().entrySet())
                 logger.error("Error connecting to " + entry.getKey(),  entry.getValue());
             throw new RuntimeException(e);
         }
@@ -82,7 +83,7 @@ public class AuthenticationTest {
                    .connect();
         } catch (NoHostAvailableException e) {
 
-            for (Map.Entry<InetAddress, Throwable> entry : e.getErrors().entrySet())
+            for (Map.Entry<InetSocketAddress, Throwable> entry : e.getErrors().entrySet())
                 logger.info("Error connecting to " + entry.getKey() + ": " + entry.getValue());
             throw new RuntimeException(e);
         }
@@ -96,7 +97,7 @@ public class AuthenticationTest {
                               .connect();
         } catch (NoHostAvailableException e) {
 
-            for (Map.Entry<InetAddress, Throwable> entry : e.getErrors().entrySet())
+            for (Map.Entry<InetSocketAddress, Throwable> entry : e.getErrors().entrySet())
                 logger.info("Error connecting to " + entry.getKey() + ": " + entry.getValue());
             throw new RuntimeException(e);
         }